backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
aiosqlite==0.19.0
psycopg2-binary==2.9.9

# Task Queue
celery==5.3.4
redis[hiredis]==5.0.1
flower==2.0.1

# Caching
aiocache==0.12.2

# LLM Clients
openai==1.3.7
anthropic==0.7.7
google-generativeai==0.3.1
httpx==0.25.2

# Vector Database
qdrant-client==1.15.1

# Object Storage
minio==7.2.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
tenacity==8.2.3
python-slugify==8.0.1
requests==2.31.0
python-magic==0.4.27

# Monitoring
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.38.0

# Distributed Tracing
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-sqlalchemy==0.42b0
opentelemetry-instrumentation-redis==0.42b0
opentelemetry-instrumentation-requests==0.42b0
opentelemetry-exporter-otlp-proto-grpc==1.21.0

# Structured Logging
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
faker==20.1.0

# Code Quality
ruff==0.1.7
mypy==1.7.1

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.2

# Other
python-dateutil==2.8.2
pytz==2023.3
